Reduce repeat of same question, Answer input validation
Repetition:
If the range is short, random will repeat the same x, y combination. To
avoid repetition, the generated number is compared to history, if the
combination (x,y or y,x) was already used, then it attempts another
random number. Given the no. of question and range, repetition can't be
avoided completely so there is a 'give up after' count (3) after which a
repeating combination will be used.

Answer input validation:
Only valid int will be accepted